update samples from Release-18 as a part of 1.1.0rc0 SDK experimental release
