--- how-to-use-azureml/work-with-data/datasets-tutorial/pipeline-with-datasets/keras-mnist-fashion/train.py
@@ -1,120 +1,0 @@
-import keras
-from keras.models import Sequential
-from keras.layers import Dense, Dropout, Flatten
-from keras.layers import Conv2D, MaxPooling2D
-from keras.layers.normalization import BatchNormalization
-from keras.utils import to_categorical
-from keras.callbacks import Callback
-
-import numpy as np
-import pandas as pd
-import os
-import matplotlib.pyplot as plt
-from sklearn.model_selection import train_test_split
-from azureml.core import Run
-
-# dataset object from the run
-run = Run.get_context()
-dataset = run.input_datasets['prepared_fashion_ds']
-
-# split dataset into train and test set
-(train_dataset, test_dataset) = dataset.random_split(percentage=0.8, seed=111)
-
-# load dataset into pandas dataframe
-data_train = train_dataset.to_pandas_dataframe()
-data_test = test_dataset.to_pandas_dataframe()
-
-img_rows, img_cols = 28, 28
-input_shape = (img_rows, img_cols, 1)
-
-X = np.array(data_train.iloc[:, 1:])
-y = to_categorical(np.array(data_train.iloc[:, 0]))
-
-# here we split validation data to optimiza classifier during training
-X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.2, random_state=13)
-
-# test data
-X_test = np.array(data_test.iloc[:, 1:])
-y_test = to_categorical(np.array(data_test.iloc[:, 0]))
-
-
-X_train = X_train.reshape(X_train.shape[0], img_rows, img_cols, 1).astype('float32') / 255
-X_test = X_test.reshape(X_test.shape[0], img_rows, img_cols, 1).astype('float32') / 255
-X_val = X_val.reshape(X_val.shape[0], img_rows, img_cols, 1).astype('float32') / 255
-
-batch_size = 256
-num_classes = 10
-epochs = 10
-
-# construct neuron network
-model = Sequential()
-model.add(Conv2D(32, kernel_size=(3, 3),
-                 activation='relu',
-                 kernel_initializer='he_normal',
-                 input_shape=input_shape))
-model.add(MaxPooling2D((2, 2)))
-model.add(Dropout(0.25))
-model.add(Conv2D(64, (3, 3), activation='relu'))
-model.add(MaxPooling2D(pool_size=(2, 2)))
-model.add(Dropout(0.25))
-model.add(Conv2D(128, (3, 3), activation='relu'))
-model.add(Dropout(0.4))
-model.add(Flatten())
-model.add(Dense(128, activation='relu'))
-model.add(Dropout(0.3))
-model.add(Dense(num_classes, activation='softmax'))
-
-model.compile(loss=keras.losses.categorical_crossentropy,
-              optimizer=keras.optimizers.Adam(),
-              metrics=['accuracy'])
-
-# start an Azure ML run
-run = Run.get_context()
-
-
-class LogRunMetrics(Callback):
-    # callback at the end of every epoch
-    def on_epoch_end(self, epoch, log):
-        # log a value repeated which creates a list
-        run.log('Loss', log['loss'])
-        run.log('Accuracy', log['accuracy'])
-
-
-history = model.fit(X_train, y_train,
-                    batch_size=batch_size,
-                    epochs=epochs,
-                    verbose=1,
-                    validation_data=(X_val, y_val),
-                    callbacks=[LogRunMetrics()])
-
-score = model.evaluate(X_test, y_test, verbose=0)
-
-# log a single value
-run.log("Final test loss", score[0])
-print('Test loss:', score[0])
-
-run.log('Final test accuracy', score[1])
-print('Test accuracy:', score[1])
-
-plt.figure(figsize=(6, 3))
-plt.title('Fashion MNIST with Keras ({} epochs)'.format(epochs), fontsize=14)
-plt.plot(history.history['accuracy'], 'b-', label='Accuracy', lw=4, alpha=0.5)
-plt.plot(history.history['loss'], 'r--', label='Loss', lw=4, alpha=0.5)
-plt.legend(fontsize=12)
-plt.grid(True)
-
-# log an image
-run.log_image('Loss v.s. Accuracy', plot=plt)
-
-# create a ./outputs/model folder in the compute target
-# files saved in the "./outputs" folder are automatically uploaded into run history
-os.makedirs('./outputs/model', exist_ok=True)
-
-# serialize NN architecture to JSON
-model_json = model.to_json()
-# save model JSON
-with open('./outputs/model/model.json', 'w') as f:
-    f.write(model_json)
-# save model weights
-model.save_weights('./outputs/model/model.h5')
-print("model saved in ./outputs/model folder")
