+++ tutorials/create-first-ml-experiment/tutorial-1st-experiment-sdk-train.yml
@@ -1,0 +1,6 @@
+name: tutorial-1st-experiment-sdk-train
+dependencies:
+- pip:
+  - azureml-sdk
+  - sklearn
+  - azureml-opendatasets
