+++ tutorials/image-classification-mnist-data/img-classification-part2-deploy.yml
@@ -1,0 +1,6 @@
+name: img-classification-part2-deploy
+dependencies:
+- pip:
+  - azureml-sdk
+  - matplotlib
+  - sklearn
