--- tutorials/img-classification-part1-training.yml
@@ -1,9 +1,0 @@
-name: img-classification-part1-training
-dependencies:
-- pip:
-  - azureml-sdk
-  - azureml-widgets
-  - matplotlib
-  - sklearn
-  - pandas
-  - azureml-opendatasets
