--- tutorials/regression-automated-ml.yml
@@ -1,7 +1,0 @@
-name: regression-automated-ml
-dependencies:
-- pip:
-  - azureml-sdk
-  - azureml-train-automl
-  - azureml-widgets
-  - azureml-opendatasets
