+++ tutorials/regression-automl-nyc-taxi-data/regression-automated-ml.yml
@@ -1,0 +1,7 @@
+name: regression-automated-ml
+dependencies:
+- pip:
+  - azureml-sdk
+  - azureml-train-automl
+  - azureml-widgets
+  - azureml-opendatasets
