--- tutorials/tutorial-1st-experiment-sdk-train.yml
@@ -1,6 +1,0 @@
-name: tutorial-1st-experiment-sdk-train
-dependencies:
-- pip:
-  - azureml-sdk
-  - sklearn
-  - azureml-opendatasets
